+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Fixed
 - Quote paths in config files. This is required for paths which have spaces in them.
 - Allow `--config-file` to be passed at any level of subcommands.
+- Allow `clifton clear-cache` to work if cache directory is missing.
 
 ### Changed
 - Retrieve OIDC client ID from the CA rather than setting it in the config. Requires version 0.3 of Conch.